requirements.txt
google-genai>=1.0.0
sentence-transformers>=2.7.0
aiohttp>=3.9.0
pydantic>=2.7.0
redis>=5.0.0
supabase>=2.10.0
structlog>=24.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
